Move linearizer.
